refactored back button into its own component, updated files
